+++ .storybook/preview.tsx
@@ -1,0 +1,25 @@
+import { CssBaseline, ThemeProvider } from '@mui/material';
+import type { Preview } from '@storybook/react';
+import { lightTheme } from '../src/styles/theme';
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+
+  decorators: [
+    Story => (
+      <ThemeProvider theme={lightTheme}>
+        <CssBaseline />
+        <Story />
+      </ThemeProvider>
+    ),
+  ],
+};
+
+export default preview;
